Lulu: Finished all endpoints. Reorganized files. Updated README
