+++ README.md
@@ -47,11 +47,20 @@ This file already has the contents needed for a Heroku deployment.
 If the `create_app` function in `app/__init__.py` is renamed or moved, the contents of this file need to change. Otherwise, we don't anticipate this file to change.
 
 ## `Endpoints`
-| Verb  | Path  | Body of Request | What it does  |
-|---|---|---|---|
-| `GET`  | `/boards`  | None | Retrieves a list of boards  |
-| `POST`  | `/boards`  | `{ title: 'titleText', owner: 'ownerText' }`  | Creates a board with title and owner informtion   |
-| `GET`  | `/boards/<board_id>/cards`  | None  | Retrieves a list of cards under one specific board   |
-| `POST`  | `/boards/<board_id>/cards`  | `{ message: 'messageText', likes_count: likesCount, board_id: boardID }` | Creates a new card   |
-| `DELETE`  | `/boards/<board_id>/cards/<card_id>`  | None  | Deletes a card |
-| `PATCH`  | `/boards/<board_id>/cards/<card_id>/like`  | `{ likes_count: likesCount }`  | Updates the like counts of a card |
+| Verb  | Path  | Body of Request | Body of Response | What it does  |
+|---|---|---|---|---|
+| `GET`  | `/boards/<board_id>`  | None | `{ "board": { "board_id": 1, "owner": "Owner 1", "title": "Title 1" } }` | Retrieves a board  |
+
+| `GET`  | `/boards`  | None | `[ { "board_id": boardId, "owner": "ownerText", "title": "titleText" } ] `| Retrieves a list of boards  |
+
+| `POST`  | `/boards`  | `{ "title": "titleText", owner: "ownerText" }`  | `{ "board": { "board_id": boardId, "owner": "ownerText", "title": "titleText" } }` | Creates a new board with title and owner informtion   |
+
+| `GET`  | `/<board_id>/cards`  | None | `[ { "board_id": boardId, "card_id": cardId, "likes_count": likesCount, "message": "messageText" }, { "board_id": boardId, "card_id": cardId, "likes_count": likesCount, "message": "messageText" } ]` | Retrieves a list of cards under one specific board   |
+
+| `GET`  | `/<board_id>/cards/<card_id>`  | None | `{ "board_id": boardId, "card_id": cardId, "likes_count": likesCount, "message": "messageText" }` | Retrieves one card under one specific board   |
+
+| `POST`  | `/boards/<board_id>/cards`  | `{ "message": "messageText" }` | `{  "board_id": boardId, "card_id": cardId, "likes_count": likesCount, "message": "messageText" }` | Creates a new card   |
+
+| `DELETE`  | `/<card_id>'`  | None | `{ "details": "Card {cardId} {messageText} successfully deleted" }` | Deletes a card |
+
+| `PATCH`  | `/<card_id>/like`  | None | `{ "board_id": boardId, "card_id": cardId, "likes_count": likesCount, "message": "messageText" }` | Updates the like counts of a card by 1|
