+++ app/__init__.py
@@ -24,10 +24,10 @@ def create_app():
     migrate.init_app(app, db)
 
 
-    from .routes import cards_bp
+    from app.routes.card_routes import cards_bp
     app.register_blueprint(cards_bp)
 
-    from .routes import boards_bp
+    from app.routes.board_routes import boards_bp
     app.register_blueprint(boards_bp)
 
     CORS(app)
